requirements.txt
discord.py[voice]==2.3.2
youtube-dl==2021.12.17
yt-dlp==2023.12.30
spotipy==2.22.1
python-dotenv==1.0.0
PyNaCl==1.5.0
aiohttp==3.9.1
asyncio==3.4.3